fix(trusty-search): dedup colocated warm-boot indexes by root_path to stop ghost duplicates (closes #860)
Legacy entries from indexes.toml carry basename-derived IDs (e.g. `trusty-tools`)
while the colocated fs scan derives full-path-sanitized IDs (e.g.
`Users_mac_workspace_trusty-tools`) via `id_from_path`. The pre-existing ID-only
dedup in `collect_colocated_entries` never matched across these two schemes, so
a phantom empty "ghost" colocated entry was registered for every legacy root on
every daemon restart.

Fix: add `known_root_paths: &HashSet<PathBuf>` to `collect_colocated_entries`;
in `restore_indexes` (start.rs) build `seen_root_paths` from canonicalized legacy
root paths alongside `seen_ids`, then pass it. The root-path check uses equality
only — NOT sub-path — to preserve correct `IndexHierarchy` behavior for
intentionally differently-scoped indexes. `create_index_handler` is unchanged.

Tests: extract inline test block to `warm_boot_tests.rs` (keeps mod.rs < 500
lines); update existing tests for the new parameter; add
`colocated_scan_deduplicates_by_root_path_against_basename_legacy_id` which
seeds `known_ids` with a basename id AND `known_root_paths` with the canonical
root, reproducing the real bug and asserting the ghost is suppressed.
